requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.2
beautifulsoup4==4.9.1
black==19.10b0
click==7.1.2
Django==3.0.6
django-autoslug==1.9.7
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-model-choices==1.0.0
django-mptt==0.11.0
django-taggit==1.3.0
djangorestframework==3.11.0
pathspec==0.8.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pytz==2020.1
regex==2020.5.14
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
